finished w/ core - need to test auto gen roster, email, matchstats.
