+++ teamwork/apps/core/helpers.py
@@ -8,12 +8,17 @@ from django.db.models import Q
 
 # Django
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.shortcuts import get_object_or_404, redirect, render
 from django.http import (HttpResponse, HttpResponseBadRequest,
                          HttpResponseRedirect)
 
 # csv
 import csv
 import codecs
+
+from teamwork.apps.projects.models import Project
+from teamwork.apps.courses.models import Course
 
 
 """
@@ -193,3 +198,52 @@ def edit_select_members(request, slug):
         return JsonResponse(data)
 
     return HttpResponse("Failure")
+
+
+@login_required
+def search(request):
+    """
+    This works but...
+
+    If a user enters multiple search terms seperated by space,
+    only the last keyword will return results
+    - andgates
+    """
+
+    page_name = "Search"
+    page_description = "Results"
+    title = "Search Results"
+
+    context = {'page_name': page_name,
+    'page_description': page_description, 'title' : title}
+
+    if request.POST.get('q'):
+        raw_keywords = request.POST.get('q')
+        keywords = []
+
+        if raw_keywords is not None:
+            if " " in raw_keywords:
+                keywords = raw_keywords.split(" ")
+            else:
+                keywords.append(raw_keywords)
+            for q in keywords:
+                user_results = User.objects.filter(
+                    Q( first_name__contains = q ) |
+                    Q( last_name__contains = q ) |
+                    Q( username__contains = q ) ).order_by('username')
+                project_results = Project.objects.filter(
+                    Q( title__contains = q ) |
+                    Q( content__contains = q ) |
+                    Q( tagline__contains = q ) ).order_by('title')
+                course_results = Course.objects.filter(
+                    Q( name__contains = q ) |
+                    Q( info__contains = q ) ).order_by('name')
+
+            if user_results:
+                context['user_results'] = user_results
+            if project_results:
+                context['project_results'] = project_results
+            if course_results:
+                context['course_results'] = course_results
+
+    return render(request, 'core/search_results.html', context)
